requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
langchain
qdrant-client
python-dotenv
google-generativeai